refactor: extract the form field
