+++ src/app/search/page.tsx
@@ -1,0 +1,5 @@
+const Page = () => {
+  return <div />;
+};
+
+export default Page;
