+++ src/components/SearchBar/SearchBarDesktop.tsx
@@ -1,20 +1,12 @@
 'use client';
 
 import { Button } from '@components/ui/button';
-import { Input } from '@components/ui/input';
-import { H3, Small } from '@components/ui/typography';
+import { H3 } from '@components/ui/typography';
 import cn from '@lib/utils';
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from '@radix-ui/react-popover';
-import { motion } from 'framer-motion';
-import { LucideIcon } from 'lucide-react';
-import Image from 'next/image';
-import Link from 'next/link';
-import { useEffect, useRef, useState } from 'react';
-import MenuItemContainer from './MenuItemContainer';
+import { useEffect, useState } from 'react';
+import ActivityField from './fields/ActivityField';
+import CategoryFieldMenu, { Category } from './fields/CategoryFieldMenu';
+import LocationFieldMenu, { Location } from './fields/LocationFieldMenu';
 
 type SearchBarDesktopProps = {
   className: string;
@@ -27,18 +19,11 @@ type SearchBarDesktopProps = {
     url: string;
     keyword: string;
   }>;
-  locations: Array<{
-    url: string;
-    text: string;
-  }>;
-  categories: Array<{
-    icon: LucideIcon;
-    url: string;
-    text: string;
-  }>;
+  locations: Location[];
+  categories: Category[];
 };
 
-const menuAnimationVariants = {
+export const menuAnimationVariants = {
   open: ({ size = 3000, locationX = 0, locationY = 0 }) => ({
     clipPath: `circle(${size}px at ${locationX}px ${locationY}px)`,
     transition: {
@@ -100,11 +85,9 @@ const SearchBarDesktop = ({
   locations,
   categories,
 }: SearchBarDesktopProps) => {
-  const [isSearchBarMenuOpen, setSearchBarMenuOpen] = useState(false);
-  const [isCategoryMenuOpen, setCategoryMenuOpen] = useState(false);
-  const [isLocationMenuOpen, setLocationMenuOpen] = useState(false);
-  const searchBarTriggerRef = useRef<HTMLButtonElement | null>(null);
-  const searchBarInputRef = useRef<HTMLInputElement | null>(null);
+  const [isSearchBarMenuOpen, setIsSearchBarMenuOpen] = useState(false);
+  const [isCategoryMenuOpen, setIsCategoryMenuOpen] = useState(false);
+  const [isLocationMenuOpen, setIsLocationMenuOpen] = useState(false);
   const isSticky = useStickyToFixed('header');
 
   return (
@@ -126,170 +109,24 @@ const SearchBarDesktop = ({
         }}
       >
         <div className="flex grow">
-          <Popover open={isSearchBarMenuOpen}>
-            <PopoverTrigger
-              ref={searchBarTriggerRef}
-              asChild
-              onClick={() => {
-                if (searchBarInputRef.current) {
-                  searchBarInputRef.current.focus();
-                }
-              }}
-            >
-              <div
-                className={cn(
-                  'mt-[1px] grow border-b border-primary py-4 hover:cursor-pointer data-[state=open]:hover:cursor-default',
-                  `${!isSearchBarMenuOpen && ' mt-0 border-t'}`
-                )}
-              >
-                <div className="space-y-2 border-x border-primary px-4">
-                  <p className="font-bold">活動</p>
-                  <Input
-                    type="text"
-                    className="h-fit w-full border-none p-0 text-base placeholder:text-primary focus-visible:ring-0 focus-visible:ring-offset-0"
-                    placeholder="搜尋關鍵字"
-                    onFocus={() => {
-                      setSearchBarMenuOpen(() => true);
-                      setCategoryMenuOpen(() => false);
-                      setLocationMenuOpen(() => false);
-                    }}
-                    onInput={() => {
-                      // TODO: debouncing method & search method
-                    }}
-                    name="keyword"
-                    ref={searchBarInputRef}
-                  />
-                </div>
-              </div>
-            </PopoverTrigger>
-            <PopoverContent
-              sticky="always"
-              side="top"
-              align="start"
-              sideOffset={0}
-              className="mx-auto min-w-[81rem]"
-              onPointerDownOutside={(e) => {
-                const open = searchBarTriggerRef.current?.contains(
-                  e.target as Node
-                );
-                setSearchBarMenuOpen(!!open);
-              }}
-              onEscapeKeyDown={() => setSearchBarMenuOpen(() => false)}
-              // prevent auto focusing for input text
-              onOpenAutoFocus={(e) => e.preventDefault()}
-              onInteractOutside={(e) => e.preventDefault()}
-            >
-              <motion.div
-                className="border-x border-t border-primary bg-surface py-12"
-                variants={menuAnimationVariants}
-                initial="closed"
-                animate={isSearchBarMenuOpen ? 'open' : 'closed'}
-                custom={{ size: 2500, locationX: 0, locationY: 500 }}
-              >
-                <p className="ml-4 text-base font-bold">推薦活動</p>
-                <div className="no-scrollbar mt-6 flex gap-4 overflow-x-auto overflow-y-hidden px-4">
-                  {activityPictures.map((item) => (
-                    <div className="min-w-fit space-y-2" key={item.description}>
-                      {/* TODO: link to search page */}
-                      <Image
-                        src={item.thumbnail}
-                        alt={item.description}
-                        width={200}
-                        height={100}
-                        className="h-[6.25rem] w-[12.5rem] object-cover"
-                      />
-                      <Small>{item.description}</Small>
-                    </div>
-                  ))}
-                </div>
-                <div className="mt-10 px-4 xl:mt-12">
-                  <p className="text-base font-bold">熱門關鍵字</p>
-                  <div className="mt-6 flex flex-wrap gap-2 overflow-x-auto overflow-y-hidden">
-                    {/* TODO: link to search page */}
-                    {activityKeywords.map((item) => (
-                      <Link
-                        href={item.url}
-                        className="w-fit rounded-2xl border px-4 py-2 font-medium"
-                        key={item.keyword}
-                      >
-                        {item.keyword}
-                      </Link>
-                    ))}
-                  </div>
-                </div>
-              </motion.div>
-            </PopoverContent>
-          </Popover>
-          <Popover onOpenChange={(e) => setCategoryMenuOpen(() => e)}>
-            <PopoverTrigger asChild>
-              <div
-                className={cn(
-                  'mt-[1px] min-w-64 border-b border-primary py-4 pl-4  hover:cursor-pointer',
-                  `${!isCategoryMenuOpen && ' mt-0 border-t'}`
-                )}
-              >
-                <button
-                  className="block w-full space-y-2 border-r border-primary px-4 text-left"
-                  type="button"
-                >
-                  <p className="font-bold">類型</p>
-                  <p className="text-base text-primary">選擇活動類型</p>
-                </button>
-              </div>
-            </PopoverTrigger>
-            <PopoverContent
-              sticky="always"
-              side="top"
-              align="start"
-              sideOffset={0}
-              className="h-[22.625rem] w-64"
-            >
-              <motion.div
-                className="absolute inset-0 border-x border-t border-primary bg-surface"
-                variants={menuAnimationVariants}
-                initial="closed"
-                animate={isCategoryMenuOpen ? 'open' : 'closed'}
-                custom={{ size: 1000, locationX: 128, locationY: 362 }}
-              >
-                <MenuItemContainer data={categories} />
-              </motion.div>
-            </PopoverContent>
-          </Popover>
-          <Popover onOpenChange={(e) => setLocationMenuOpen(() => e)}>
-            <PopoverTrigger asChild>
-              <div
-                className={cn(
-                  'mt-[1px] min-w-64 border-b border-primary py-4 pl-4 hover:cursor-pointer',
-                  `${!isLocationMenuOpen && ' mt-0 border-t'}`
-                )}
-              >
-                <button
-                  className="block w-full space-y-2 border-primary px-4 text-left"
-                  type="button"
-                >
-                  <p className="font-bold">地區</p>
-                  <p className="text-base text-primary">選擇活動地區</p>
-                </button>
-              </div>
-            </PopoverTrigger>
-            <PopoverContent
-              sticky="always"
-              side="top"
-              align="start"
-              sideOffset={0}
-              className="relative h-[22.625rem] w-64"
-            >
-              <motion.div
-                className="absolute inset-0 border-x border-t border-primary bg-surface"
-                variants={menuAnimationVariants}
-                initial="closed"
-                animate={isLocationMenuOpen ? 'open' : 'closed'}
-                custom={{ size: 1000, locationX: 128, locationY: 362 }}
-              >
-                <MenuItemContainer data={locations} />
-              </motion.div>
-            </PopoverContent>
-          </Popover>
+          <ActivityField
+            activityKeywords={activityKeywords}
+            activityPictures={activityPictures}
+            isSearchBarMenuOpen={isSearchBarMenuOpen}
+            setIsCategoryMenuOpen={setIsCategoryMenuOpen}
+            setIsLocationMenuOpen={setIsLocationMenuOpen}
+            setIsSearchBarMenuOpen={setIsSearchBarMenuOpen}
+          />
+          <CategoryFieldMenu
+            isCategoryMenuOpen={isCategoryMenuOpen}
+            setIsCategoryMenuOpen={setIsCategoryMenuOpen}
+            categories={categories}
+          />
+          <LocationFieldMenu
+            isLocationMenuOpen={isLocationMenuOpen}
+            setIsLocationMenuOpen={setIsLocationMenuOpen}
+            locations={locations}
+          />
         </div>
         <Button className="flex h-auto self-auto px-20 text-xl font-bold">
           搜尋活動
@@ -298,5 +135,4 @@ const SearchBarDesktop = ({
     </section>
   );
 };
-
 export default SearchBarDesktop;
