+++ src/components/SearchBar/fields/ActivityField.tsx
@@ -1,0 +1,140 @@
+'use client';
+
+import { Input } from '@components/ui/input';
+import { Small } from '@components/ui/typography';
+import cn from '@lib/utils';
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from '@radix-ui/react-popover';
+import { motion } from 'framer-motion';
+import Image from 'next/image';
+import Link from 'next/link';
+import { Dispatch, SetStateAction, useRef } from 'react';
+import menuAnimationVariants from './utils';
+
+type ActivityFieldProps = {
+  activityKeywords: {
+    url: string;
+    keyword: string;
+  }[];
+  activityPictures: {
+    thumbnail: string;
+    url: string;
+    description: string;
+  }[];
+  isSearchBarMenuOpen: boolean;
+  setIsCategoryMenuOpen: Dispatch<SetStateAction<boolean>>;
+  setIsLocationMenuOpen: Dispatch<SetStateAction<boolean>>;
+  setIsSearchBarMenuOpen: Dispatch<SetStateAction<boolean>>;
+};
+
+const ActivityField = ({
+  activityKeywords,
+  activityPictures,
+  isSearchBarMenuOpen,
+  setIsCategoryMenuOpen,
+  setIsLocationMenuOpen,
+  setIsSearchBarMenuOpen,
+}: ActivityFieldProps) => {
+  const searchBarTriggerRef = useRef<HTMLButtonElement | null>(null);
+  const searchBarInputRef = useRef<HTMLInputElement | null>(null);
+
+  return (
+    <Popover open={isSearchBarMenuOpen}>
+      <PopoverTrigger
+        ref={searchBarTriggerRef}
+        asChild
+        onClick={() => {
+          if (searchBarInputRef.current) {
+            searchBarInputRef.current.focus();
+          }
+        }}
+      >
+        <div
+          className={cn(
+            'mt-[1px] grow border-b border-primary py-4 hover:cursor-pointer data-[state=open]:hover:cursor-default',
+            `${!isSearchBarMenuOpen && ' mt-0 border-t'}`
+          )}
+        >
+          <div className="space-y-2 border-x border-primary px-4">
+            <p className="font-bold">活動</p>
+            <Input
+              type="text"
+              className="h-fit w-full border-none p-0 text-base placeholder:text-primary focus-visible:ring-0 focus-visible:ring-offset-0"
+              placeholder="搜尋關鍵字"
+              onFocus={() => {
+                setIsSearchBarMenuOpen(() => true);
+                setIsCategoryMenuOpen(() => false);
+                setIsLocationMenuOpen(() => false);
+              }}
+              onInput={() => {
+                // TODO: debouncing method & search method
+              }}
+              name="keyword"
+              ref={searchBarInputRef}
+            />
+          </div>
+        </div>
+      </PopoverTrigger>
+      <PopoverContent
+        sticky="always"
+        side="top"
+        align="start"
+        sideOffset={0}
+        className="mx-auto min-w-[81rem]"
+        onPointerDownOutside={(e) => {
+          const open = searchBarTriggerRef.current?.contains(e.target as Node);
+          setIsSearchBarMenuOpen(!!open);
+        }}
+        onEscapeKeyDown={() => setIsSearchBarMenuOpen(() => false)}
+        // prevent auto focusing for input text
+        onOpenAutoFocus={(e) => e.preventDefault()}
+        onInteractOutside={(e) => e.preventDefault()}
+      >
+        <motion.div
+          className="border-x border-t border-primary bg-surface py-12"
+          variants={menuAnimationVariants}
+          initial="closed"
+          animate={isSearchBarMenuOpen ? 'open' : 'closed'}
+          custom={{ size: 2500, locationX: 0, locationY: 500 }}
+        >
+          <p className="ml-4 text-base font-bold">推薦活動</p>
+          <div className="no-scrollbar mt-6 flex gap-4 overflow-x-auto overflow-y-hidden px-4">
+            {activityPictures.map((item) => (
+              <div className="min-w-fit space-y-2" key={item.description}>
+                {/* TODO: link to search page */}
+                <Image
+                  src={item.thumbnail}
+                  alt={item.description}
+                  width={200}
+                  height={100}
+                  className="h-[6.25rem] w-[12.5rem] object-cover"
+                />
+                <Small>{item.description}</Small>
+              </div>
+            ))}
+          </div>
+          <div className="mt-10 px-4 xl:mt-12">
+            <p className="text-base font-bold">熱門關鍵字</p>
+            <div className="mt-6 flex flex-wrap gap-2 overflow-x-auto overflow-y-hidden">
+              {/* TODO: link to search page */}
+              {activityKeywords.map((item) => (
+                <Link
+                  href={item.url}
+                  className="w-fit rounded-2xl border px-4 py-2 font-medium"
+                  key={item.keyword}
+                >
+                  {item.keyword}
+                </Link>
+              ))}
+            </div>
+          </div>
+        </motion.div>
+      </PopoverContent>
+    </Popover>
+  );
+};
+
+export default ActivityField;
