+++ src/components/SearchBar/fields/utils.ts
@@ -1,0 +1,21 @@
+const menuAnimationVariants = {
+  open: ({ size = 3000, locationX = 0, locationY = 0 }) => ({
+    clipPath: `circle(${size}px at ${locationX}px ${locationY}px)`,
+    transition: {
+      type: 'spring',
+      stiffness: 30,
+      restDelta: 2,
+    },
+  }),
+  closed: ({ locationX = 0, locationY = 0 }) => ({
+    clipPath: `circle(0px at ${locationX}px ${locationY}px)`,
+    transition: {
+      delay: 0.5,
+      type: 'spring',
+      stiffness: 400,
+      damping: 40,
+    },
+  }),
+};
+
+export default menuAnimationVariants;
